Prepare for adding new version of using Complex = typename boost::multiprecision::make_boost_complex<Real>::type;
