+++ include/boost/math/fft/fftw_backend.hpp
@@ -13,7 +13,7 @@
 
   #include <fftw3.h>
   #include <boost/math/fft/dft_api.hpp>
-  #include <boost/math/fft/abstract_ring.hpp>
+  #include <boost/math/fft/multiprecision_complex.hpp>
 
   namespace boost { namespace math {  namespace fft {
