+++ test/fft_correctedness.cpp
@@ -1,9 +1,9 @@
 #include "math_unit_test.hpp"
+#include <boost/math/fft/multiprecision_complex.hpp>
 #include <boost/math/fft/bsl_backend.hpp>
 #include <boost/math/fft/fftw_backend.hpp>
 #include <boost/math/fft/gsl_backend.hpp>
 #include <boost/math/fft/algorithms.hpp>
-#include <boost/math/fft/abstract_ring.hpp>
 #include <boost/math/constants/constants.hpp>
 #ifdef BOOST_MATH_USE_FLOAT128
 #include <boost/multiprecision/complex128.hpp>
